# Conflicts:
#   spec/tasks/gitlab/backup_rake_spec.rb

+++ app/services/milestones/destroy_service.rb
@@ -1,10 +1,10 @@
+# frozen_string_literal: true
+
 module Milestones
   class DestroyService < Milestones::BaseService
     def execute(milestone)
-      return unless milestone.project_milestone?
-
       Milestone.transaction do
-        update_params = { milestone: nil }
+        update_params = { milestone: nil, skip_milestone_email: true }
 
         milestone.issues.each do |issue|
           Issues::UpdateService.new(parent, current_user, update_params).execute(issue)
@@ -14,14 +14,20 @@ module Milestones
           MergeRequests::UpdateService.new(parent, current_user, update_params).execute(merge_request)
         end
 
-        event_service.destroy_milestone(milestone, current_user)
-
-        Event.for_milestone_id(milestone.id).each do |event|
-          event.target_id = nil
-          event.save
-        end
+        log_destroy_event_for(milestone)
 
         milestone.destroy
+      end
+    end
+
+    def log_destroy_event_for(milestone)
+      return if milestone.group_milestone?
+
+      event_service.destroy_milestone(milestone, current_user)
+
+      Event.for_milestone_id(milestone.id).each do |event|
+        event.target_id = nil
+        event.save
       end
     end
   end
